requirements.txt
spotipy>=2.23.0
Pillow>=10.0.0
pygame>=2.5.0
python-evdev>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
